Fixed segment example
